feat: reduced i18n boilerplate
